// src/app/(home)/_config/skills/programs/enums.ts
export enum HOBBY_TYPE {
  DESIGN = "Design",
  PROGRAMMING = "Programming",
  PHOTOGRAPHY = "Photography",
  ART = "Art",
  MUSIC = "Music",
  MODELING = "Modeling",
}
export enum PROGRAMM_TYPE {
  SITES = "Sites",
  DOCUMENTS = "Documents",
  DAW = "DAW",
  VIDEO = "Video",
  STREAMING = "Streaming",
  SUBTITLES = "Subtitles",
  IDE = "IDE",
  GAME_ENGINE = "Game Engine",
  DEVELOPMENT_TOOLS = "Development Tools",
}